+++ publish/binary-compiler/src/entry.ts
@@ -1,54 +1,10 @@
+// setup plugin
+import "./secretlint-plugin.ts";
 // preload for embedded binary
-import * as preset from "@secretlint/secretlint-rule-preset-recommend";
-import * as pattern from "@secretlint/secretlint-rule-pattern";
-import * as sarif from "@secretlint/secretlint-formatter-sarif";
 import { cli, run } from "secretlint/cli";
-import { plugin, type PluginBuilder } from "bun";
 import * as fs from "node:fs";
 import * as path from "node:path";
 
-const mocks = {
-    "@secretlint/secretlint-rule-preset-recommend": preset,
-    "@secretlint/secretlint-rule-pattern": pattern,
-    "@secretlint/secretlint-formatter-sarif": sarif
-};
-const mockNames = Object.keys(mocks);
-await plugin({
-    name: "secretlint",
-    setup(build: PluginBuilder): void | Promise<void> {
-        // require.resolve hooks
-        build.onResolve({ filter: /@secretlint/ }, (args) => {
-            // if match the path with mocks, return mock name
-            const match = args.path.match(/(@secretlint\/[^/]*)/);
-            if (match && mockNames.includes(match[1])) {
-                // return `secretlint:${name}` instead of file path
-                return {
-                    path: match[1],
-                    namespace: "secretlint"
-                };
-            }
-            // if not match, return original path
-            return args;
-        });
-        // require/import/dynamic import hooks
-        Object.entries(mocks).forEach(([name, mock]) => {
-            // build.module will return mock object
-            build.module(`${name}`, () => {
-                return {
-                    exports: mock,
-                    loader: "object"
-                };
-            });
-            // build.onResolve will return `secretlint:${name}` instead of file path
-            build.module(`secretlint:${name}`, () => {
-                return {
-                    exports: mock,
-                    loader: "object"
-                };
-            });
-        });
-    }
-});
 // --init override
 if (cli.flags.init) {
     // write .secretlintrc.json
@@ -72,8 +28,6 @@ if (cli.flags.init) {
     console.log("Create .secretlintrc.json");
     process.exit(0);
 }
-// FIXME: wait to start for hook plugin
-await new Promise((resolve) => setTimeout(resolve, 100));
 // secretlint CLI wrapper
 // TODO: --version does not work because package.json is not bundled
 run(cli.input, cli.flags).then(
